requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
streamlit==1.28.1
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1
xgboost==2.0.2
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
pdfplumber==0.10.3
aiofiles==23.2.1
Pillow==10.1.0
opencv-python==4.8.1.78
transformers==4.35.2
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
